app/requirements.txt
# requirements.txt

flask
psycopg2-binary
sqlalchemy
python-dotenv